backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==13.0

# Database
sqlalchemy==2.0.34
aiosqlite==0.20.0
asyncpg==0.29.0
alembic==1.13.0

# Data Sources
ccxt==4.4.0
yfinance==0.2.40

# Data Processing
pandas==2.2.2
numpy==1.26.4
pandas-ta==0.3.14b1

# API
pydantic==2.9.0
python-dotenv==1.0.1
httpx==0.27.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
